+++ src/test/scala/sangria/schema/TypeFieldUniquenessSpec.scala
@@ -6,13 +6,13 @@ class TypeFieldUniquenessSpec extends WordSpec with Matchers {
 
   "ObjectType" should {
     "allow unique fields" in {
-      ObjectType("Test", List[Field[Unit, Unit]](
+      ObjectType("Test", fields[Unit, Unit](
         Field("a", StringType, resolve = _ => "foo"),
         Field("b", StringType, resolve = _ => "foo"),
         Field("c", StringType, resolve = _ => "foo")
       ))
 
-      ObjectType("Test", () => List[Field[Unit, Unit]](
+      ObjectType("Test", () => fields[Unit, Unit](
         Field("a", StringType, resolve = _ => "foo"),
         Field("b", StringType, resolve = _ => "foo"),
         Field("c", StringType, resolve = _ => "foo")
@@ -21,7 +21,7 @@ class TypeFieldUniquenessSpec extends WordSpec with Matchers {
 
     "disallow non-unique fields" in {
       an [IllegalArgumentException] should be thrownBy {
-        ObjectType("Test", List[Field[Unit, Unit]](
+        ObjectType("Test", fields[Unit, Unit](
           Field("a", StringType, resolve = _ => "foo"),
           Field("b", StringType, resolve = _ => "foo"),
           Field("a", StringType, resolve = _ => "foo")
@@ -29,7 +29,7 @@ class TypeFieldUniquenessSpec extends WordSpec with Matchers {
       }
 
       an [IllegalArgumentException] should be thrownBy {
-        ObjectType("Test", () => List[Field[Unit, Unit]](
+        ObjectType("Test", () => fields[Unit, Unit](
           Field("a", StringType, resolve = _ => "foo"),
           Field("b", StringType, resolve = _ => "foo"),
           Field("a", StringType, resolve = _ => "foo")
@@ -40,13 +40,13 @@ class TypeFieldUniquenessSpec extends WordSpec with Matchers {
 
   "InterfaceType" should {
     "allow unique fields" in {
-      InterfaceType("Test", List[Field[Unit, Unit]](
+      InterfaceType("Test", fields[Unit, Unit](
         Field("a", StringType, resolve = _ => "foo"),
         Field("b", StringType, resolve = _ => "foo"),
         Field("c", StringType, resolve = _ => "foo")
       ))
 
-      InterfaceType("Test", () => List[Field[Unit, Unit]](
+      InterfaceType("Test", () => fields[Unit, Unit](
         Field("a", StringType, resolve = _ => "foo"),
         Field("b", StringType, resolve = _ => "foo"),
         Field("c", StringType, resolve = _ => "foo")
@@ -55,7 +55,7 @@ class TypeFieldUniquenessSpec extends WordSpec with Matchers {
 
     "disallow non-unique fields" in {
       an [IllegalArgumentException] should be thrownBy {
-        InterfaceType("Test", List[Field[Unit, Unit]](
+        InterfaceType("Test", fields[Unit, Unit](
           Field("a", StringType, resolve = _ => "foo"),
           Field("b", StringType, resolve = _ => "foo"),
           Field("a", StringType, resolve = _ => "foo")
@@ -63,7 +63,7 @@ class TypeFieldUniquenessSpec extends WordSpec with Matchers {
       }
 
       an [IllegalArgumentException] should be thrownBy {
-        InterfaceType("Test", () => List[Field[Unit, Unit]](
+        InterfaceType("Test", () => fields[Unit, Unit](
           Field("a", StringType, resolve = _ => "foo"),
           Field("b", StringType, resolve = _ => "foo"),
           Field("a", StringType, resolve = _ => "foo")